% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm implementation}
\usage{
dijkstra(mat, weights)
}
\arguments{
\item{mat}{the Graph structure representation as output from as_edgelist(G)}

\item{weights}{a vector containing the weights for each edge(each line) of
the edgelist matrix}
}
\description{
Dijkstra algorithm implementation
}
