% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstraSparseMatrix}
\alias{dijkstraSparseMatrix}
\title{Dijkstra algorithm implementation - input a SparseMatrix}
\usage{
dijkstraSparseMatrix(dgCMatrix, startNode)
}
\arguments{
\item{dgCMatrix}{the Graph structure representation as output from
as_adj(graph, attr = "weight")}

\item{startNode}{the label of the starting node which we are interested in +
calculating the shortest paths}
}
\description{
Dijkstra algorithm implementation - input a SparseMatrix
}
