% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm implementation}
\usage{
dijkstra(edgeList, weights, numNodes, startNode, isDirected)
}
\arguments{
\item{edgeList}{the Graph structure representation as output from as_edgelist(G)}

\item{weights}{a vector containing the weights for each edge(each line) of
the edgelist matrix}

\item{numNodes}{the number of nodes of the graph}

\item{startNode}{the starting node which we are interested in calculating
the shortest paths}

\item{isDirected}{boolean value stating that each edge into the edge matrix
should be trated or not as a directed edge}
}
\description{
Dijkstra algorithm implementation
}
